+++ src/server.py
@@ -1,8 +1,5 @@
 import random
 
-from mesa import Agent, Model
-from mesa.space import MultiGrid
-from mesa.time import RandomActivation
 from mesa.visualization.ModularVisualization import ModularServer
 from mesa.visualization.modules import CanvasGrid
 #from mesa_models.boltzmann_wealth_model.model import BoltzmannWealthModel
@@ -11,88 +8,15 @@ import ipywidgets as widgets
 from IPython.display import display
 from mesa.visualization.UserParam import Slider
 
+from model import *
+from agent import *
+
 # Defina suas variáveis iniciais aqui
-
-# Quando instanciar a classe, passa o list(x,y) da origem do avião e o list(x,y) do aeroporto em que ele vai viajar
-class Airplane(Agent):
-    def __init__(self, unique_id, model,origin : list ,destination : list):
-        super().__init__(unique_id, model)
-        self.destination = destination
-        self.origin = origin
-
-    def step(self):
-        self.move()
-
-    def move(self):
-        x0 = self.origin[0]
-        y0 = self.origin[1]
-        x1 = self.destination[0]
-        y1 = self.destination[1]
-
-        if [x0,y0] != [x1,y1]:
-            if x0 < x1:
-                x0 = x0 + 1
-                self.model.grid.move_agent(self, (x0,y0))
-            elif x0 > x1 :
-                x0 = x0 - 1
-                self.model.grid.move_agent(self, (x0,y0))
-            if y0 < y1:
-                y0 = y0 + 1
-                self.model.grid.move_agent(self, (x0,y0))
-            elif y0 > y1:
-                y0 = y0 - 1
-                self.model.grid.move_agent(self, (x0,y0))
-        
-        self.origin[0] = x0
-        self.origin[1] = y0
-        self.destination[0] = x1
-        self.destination[1] = y1
-
-
-class Airport(Agent):
-    def __init__(self, unique_id, model, pos):
-        super().__init__(unique_id, model)
-        self.pos = pos
-
-    def step(self):
-        # Aeroportos são agentes fixos, não têm movimento
-        pass
-
-
 def agent_portrayal(agent):
     if isinstance(agent, Airplane):
         return {"Shape": "circle", "Filled": "true", "Layer": 0, "Color": "red", "r": 0.5}
     elif isinstance(agent, Airport):
         return {"Shape": "rect", "Filled": "true", "Layer": 0, "Color": "blue", "w": 0.5, "h": 0.5}
-
-
-class TraficControler(Model):
-    def __init__(self, airplanes, width, height, airports):
-        self.airplanes = airplanes
-        self.airports = airports
-        self.grid = MultiGrid(width, height, True)
-        self.schedule = RandomActivation(self)        
-        self.airport_coordinates = []
-        
-        for i in range(self.airports):
-            x = self.random.randrange(self.grid.width)
-            y = self.random.randrange(self.grid.height)
-            self.airport_coordinates.append((x,y))
-            airport = Airport(i, self, (x, y))
-            self.grid.place_agent(airport, (x, y))
-        
-        total = len(self.airport_coordinates) # Obtem a quantidade de aeroportos
-        
-        # Criação dos agentes móveis
-        for i in range(self.airplanes):
-            dest = list(self.airport_coordinates[(i+1)%total])
-            origin = list(self.airport_coordinates[i%total])
-            a = Airplane(i, self,origin,dest)
-            self.schedule.add(a)
-            self.grid.place_agent(a, origin)
-
-    def step(self):
-        self.schedule.step()
 
 grid = CanvasGrid(agent_portrayal, 20, 20, 500, 500)
